added authentication functionality in journal entry controller

+++ myJournal/src/main/java/JournalApp/myJournal/controller/JournalEntryControllerV2.java
@@ -8,11 +8,14 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/journal")
@@ -24,49 +27,75 @@ public class JournalEntryControllerV2 {
     @Autowired
     private UserService userService;
 
-        @GetMapping("/{username}")
-        public ResponseEntity<?> getEntriesByUser(@PathVariable String username) {
+        @GetMapping
+        public ResponseEntity<?> getEntriesByUser() {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String username = authentication.getName();
             User user = userService.findByUserName(username);
             List<JournalEntry> all = user.getJournalEntries();
             if(all != null && !all.isEmpty()){
-                return new ResponseEntity<>(HttpStatus.OK);
+                return new ResponseEntity<>(all, HttpStatus.OK);
             }
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
 
-        @PostMapping("/{username}")
-        public ResponseEntity<JournalEntry> createEntry(@RequestBody JournalEntry myEntry, @PathVariable String username) {
-            journalEntryService.saveEntry(myEntry, username);
-            return new ResponseEntity<>(HttpStatus.CREATED);
+
+
+        @PostMapping
+        public ResponseEntity<?> createEntry(@RequestBody JournalEntry myEntry) {
+            try {
+                Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+                String username = authentication.getName();
+                journalEntryService.saveEntry(myEntry, username);
+                return new ResponseEntity<>(HttpStatus.CREATED);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }
 
-        @GetMapping("/id/{myID}")
-        public ResponseEntity<JournalEntry> findViaId(@PathVariable ObjectId myID){
-            Optional<JournalEntry> journalEntry = journalEntryService.findById(myID);
-            if(journalEntry.isPresent()){
-                return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
+        @GetMapping("/id/{myId}")
+        public ResponseEntity<JournalEntry> findViaId(@PathVariable ObjectId myId){
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String username = authentication.getName();
+            User user = userService.findByUserName(username);
+            List<JournalEntry> collect = user.getJournalEntries().stream().filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+            if(!collect.isEmpty()) {
+                Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
+                if(journalEntry.isPresent()) return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
             }
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
-        @DeleteMapping("/id/{username}/{myId}")
-        public ResponseEntity<JournalEntry> deleteEntry(@PathVariable String username, @PathVariable ObjectId myId) {
-            Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
-            if(journalEntry.isPresent()){
-                journalEntryService.deleteById(myId, username);
-                return new ResponseEntity<>(HttpStatus.OK);
+        @DeleteMapping("/id/{myId}")
+        public ResponseEntity<JournalEntry> deleteEntry(@PathVariable ObjectId myId) {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String username = authentication.getName();User user = userService.findByUserName(username);
+            List<JournalEntry> collect = user.getJournalEntries().stream().filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+            if(!collect.isEmpty()){
+                Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
+                if(journalEntry.isPresent()) {
+                    journalEntryService.deleteById(myId, username);
+                    return new ResponseEntity<>(HttpStatus.OK);
+                }
             }
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
         @PutMapping("id/{myId}")
         public ResponseEntity<JournalEntry> updateEntry( @PathVariable ObjectId myId, @RequestBody JournalEntry newEntry){
-            JournalEntry old = journalEntryService.findById(myId).orElse(null);
-            if(old != null){
-                old.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().equals("") ? newEntry.getTitle() : old.getTitle());
-                old.setContent(newEntry.getContent() != null && !newEntry.getContent().equals("") ? newEntry.getContent() : old.getContent());
-                //journalEntryService.saveEntry(old);
-                return new ResponseEntity<>(old, HttpStatus.OK );
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String username = authentication.getName();
+            User user = userService.findByUserName(username);
+            List<JournalEntry> collect = user.getJournalEntries().stream().filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+            if(!collect.isEmpty()) {
+                Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
+                if(journalEntry.isPresent()) {
+                    JournalEntry old = journalEntry.get();
+                    old.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().equals("") ? newEntry.getTitle() : old.getTitle());
+                    old.setContent(newEntry.getContent() != null && !newEntry.getContent().equals("") ? newEntry.getContent() : old.getContent());
+                    journalEntryService.saveEntry(old, username);
+                    return new ResponseEntity<>(old, HttpStatus.OK );
+                }
             }
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
